Added a stylesheet and changed the background colour.
I had issues with the style sheet as it was appearing in Chinese characters for a long time. I was able to resolve this through opening the files in notepad and resaving them. I found that the index file had been saved in a language that was not UTF-8, which was causing the issues.

Through solving this issue I was able to ensure my web page worked correctly and linked to the css.